+++ test/support/kinetic_ecto_test_repo.ex
@@ -2,5 +2,5 @@ defmodule KineticEcto.TestRepo do
   @moduledoc false
 
   use Ecto.Repo, otp_app: :kinetic_ecto, adapter: Ecto.Adapters.SQLite3
-  use KineticEcto.Transact
+  use KineticEcto.RepoTransact
 end