+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # KineticEcto Changelog
 
+## 1.1.1 / 2025-02-03
+
+- Renamed `KineticEcto.Transact` to `KineticEcto.RepoTransact`.
+  `use KineticEcto.Transact` will now output a warning indicating that it has
+  been replaced with `use KineticEcto.RepoTransact`. The functionality otherwise
+  remains the same.
+
 ## 1.1.0 / 2025-01-29
 
 - Added `KineticEcto.Transact`, an easy way to add Saša Juriċ's